Moved content to CardFilter
